useMovieDetails movet to other file
