+++ src/features/movies/MovieDetails/index.js
@@ -6,39 +6,13 @@ import { useSelector } from "react-redux";
 import { selectMoviesState } from "../movieSlice";
 import { Error } from "../../../common/Error";
 import { GenresContainer, GenreTag } from "../movieList/styled";
+import { useMovieDetails } from "../useMovieDetails";
 
-const useMovieDetails = () => {
-  const [movieDetails, setMovieDetails] = useState([]);
-  const [movieGenres, setMovieGenres] = useState([]);
-  const { id } = useParams();
-  const getMoviesDetails = async () => {
 
-    try {
-      const response = await fetch(`https://api.themoviedb.org/3/movie/${id}?api_key=46d96cb40fd666dc8da8fdc02b4c2019`)
-
-      if (!response.ok) {
-        throw new Error(response.statusText)
-      }
-      const movieData = await response.json()
-      const genres = movieData.genres
-
-      setMovieDetails(movieData)
-      setMovieGenres(genres)
-    }
-    catch (error) {
-      console.log(error)
-      throw new Error(error.message);
-    }
-  };
-  useEffect(() => { getMoviesDetails() }, [])
-
-  return { movieDetails, movieGenres }
-};
 
 export const MovieDetails = () => {
   const { movieDetails, movieGenres } = useMovieDetails([])
   const { loading, error } = useSelector(selectMoviesState)
-  const { id } = useParams()
 
   return (
     <>
